pyproject.toml
[tool.poetry]
name = "dna-api"
version = "0.1.0"
description = ""
authors = ["Affonso Brian Pereira Azevedo <contato.affonsobrian@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytz = "2021.1"
Django = "3.2.7"
django-configurations = "2.2"
gunicorn = "20.1.0"
newrelic = "6.8.1.164"
psycopg2-binary = "2.9.1"
dj-database-url = "0.5.0"
django-model-utils = "4.1.1"
django-unique-upload = "0.2.1"
djangorestframework = "3.12.4"
Markdown = "3.3.4"
django-filter = "2.4.0"
ipdb = "0.13.9"
ipython = "7.27.0"
mkdocs = "1.2.2"
flake8 = "3.9.2"
mock = "4.0.3"
factory-boy = "3.2.0"
django-nose = "1.4.7"
nose-progressive = "1.5.2"
coverage = "5.5"
django-storages = "1.11.1"
boto3 = "1.18.36"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
